Extract shared FpML normalization boundary
